+++ redux/actions/actions.js
@@ -1,0 +1,8 @@
+import types from './types';
+
+export default {
+  repo: {
+    setTree: data => ({ type: types.repo.SET_TREE, data }),
+    setRepo: data => ({ type: types.repo.SET_REPO, data }),
+  },
+};