+++ redux/actions/types.js
@@ -1,0 +1,7 @@
+export default {
+  repo: {
+    FETCH_TREE: 'FETCH_TREE',
+    SET_TREE: 'SET_TREE',
+    SET_REPO: 'SET_REPO',
+  },
+};