+++ redux/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import repoReducer from './repoReducer';
+
+export default combineReducers({
+  repo: repoReducer,
+});