+++ redux/sagas/index.js
@@ -1,0 +1,8 @@
+import { all } from 'redux-saga/effects';
+import apiSaga from './apiSaga';
+
+export default function* () {
+  yield all([
+    apiSaga(),
+  ]);
+}